Fix Windows-hosted libgpr builds for cross-Linux
!46 broke this configuration by introducing a scenario variable that selects
target-specific source files, and setting that variable from the "OS"
external, defaulting to "UNIX".

This works because the environment variable is only ever set on Windows…
except when targeting cross-Linux.

Set the external value explicitly for all Linux targets: it does not hurt
for native Linux.

TN: eng/gpr/gpr-issues#102