Fix out-of-tree build when using relative path with ".."
In this case when end up with full path directory like:

   /some/path/prj/src/../obj

And since GPRbuild checks and creates all directories in the path,
the directories:

   prj
   prj/src
   prj/obj

are created, but prj/src is actually not needed/used.

This is fixed by ensuring that the paths to object/exec/lib directories
are properly normalized.

TN eng/gpr/gpr-issues#212