GPRbuild better handles GNU make's jobserver
Deleted the specific switch to activate the functionality and instead added a switch
to auto-detect if the functionality can be used.
GPRbuild will warn users if it was not invoked recursively in a Makefile and GNU make's jobserver information are available.
GPRbuild correctly stores mapping files when failing to get a token in order to re-use it

Issue: eng/gpr/gpr-issues#316

Revert "Revert "GPRbuild better handles GNU make's jobserver""

This reverts commit bf3fc43729a013f8c45938eff5554a256d363f69.