Remove use of System.Os_Lib.time_t type
The type is misleading since it has nothing to do with C's time_t (it is
defined as a subtype of Long_Long_Integer) and will eventually be removed.

The change also replaces a call to Ada.Calendar.Conversions.To_Unix_Time,
which is obsolescent, by one to Ada.Calendar.Conversions.To_Unix_Time_64.

Issue: eng/toolchain/gnat#1095